+++ react/typings/vtex.responsive-values.d.ts
@@ -1,0 +1,11 @@
+declare module 'vtex.responsive-values' {
+  interface ResponsiveValue<T> {
+    tablet?: T
+    phone?: T
+    desktop?: T
+    mobile?: T
+  }
+
+  type MaybeResponsiveValue<T> = ResponsiveValue<T> | T
+  const useResponsiveValue: <T>(value: MaybeResponsiveValue<T>) => T
+}
